requirements.txt
ha_mqtt_discoverable>0.8
PyYAML
requests